+++ api/pages_queries.go
@@ -53,10 +53,6 @@ func (res *Resolver) Page(
 	}, nil
 }
 
-type PageSort struct {
-	Field string
-	Order string
-}
 type PagesArgs struct {
 	Sort *PageSort
 }
@@ -65,51 +61,5 @@ func (res *Resolver) Pages(
 	ctx context.Context,
 	args PagesArgs,
 ) (*[]*pageResolver, error) {
-	pages := []core.Page{}
-
-	query := res.
-		dataSource.
-		DB().
-		Model(&pages)
-	query = sortPages(query, args.Sort)
-	err := query.Select()
-
-	if err != nil {
-		return nil, err
-	}
-
-	resolvers := make([]*pageResolver, len(pages))
-	for i := range pages {
-		resolvers[i] = &pageResolver{
-			dataSource: res.dataSource,
-			data:       &pages[i],
-		}
-	}
-
-	return &resolvers, nil
-}
-
-func sortPages(query *orm.Query, sort *PageSort) *orm.Query {
-	if sort != nil {
-		switch sort.Field {
-		case "AUTHOR":
-			query = query.Relation("Author")
-			return query.OrderExpr("Author.email " + sort.Order)
-
-		case "IS_PUBLISHED":
-			return query.OrderExpr("is_published " + sort.Order)
-
-		case "NAME":
-			return query.OrderExpr("name " + sort.Order)
-
-		case "SLUG":
-			return query.OrderExpr("slug " + sort.Order)
-
-		case "UPDATED_AT":
-			return query.OrderExpr("updated_at " + sort.Order)
-		}
-
-	}
-
-	return query.OrderExpr("created_at ASC")
+	return resolvePages(res.dataSource, args.Sort, nil)
 }
