+++ api/schema/type/Directory.graphql
@@ -6,5 +6,5 @@ type Directory {
   parent: Directory
   isPublished: Boolean!
   children: [Directory]
-  pages: [Page]
+  pages(sort: PageSort): [Page!]
 }
